Move models to package
