+++ service/api/models/track.py
@@ -48,18 +48,3 @@ class Track(models.Model):
     @property
     def spotify(self):
         return "{}{}/{}".format(settings.DSP_BASE, self.id, "spotify")
-
-
-class Playlist(models.Model):
-    id = models.AutoField(primary_key=True)
-    title = models.CharField(max_length=200, null=False)
-    tracks = models.ManyToManyField(Track, through="PlaylistTrack", related_name="playlists")
-
-
-class PlaylistTrack(models.Model):
-    track = models.ForeignKey(Track, on_delete=models.CASCADE)
-    playlist = models.ForeignKey(Playlist, on_delete=models.CASCADE)
-    order = models.IntegerField()
-
-    class Meta:
-        unique_together = ('playlist_id', 'track_id')
